Still cannot get argparse version to work
